move flux setup
