+++ .github/workflows/deploy.yml
@@ -20,7 +20,7 @@ jobs:
     - name: checkout
       uses: actions/checkout@v2
     - name: Fake Install flux
-      run: mkdir -p cmd/wego/flux/bin && touch cmd/wego/flux/bin/flux
+      run: mkdir -p pkg/flux/bin && touch pkg/flux/bin/flux
     - name: build
       run: make all BINARY_NAME=wego-${{matrix.os}}
     - name: publish
