+++ .github/workflows/docker_scan.yaml
@@ -49,7 +49,7 @@ jobs:
       - name: Checkout
         uses: actions/checkout@v2
       - name: Fake Install flux
-        run: mkdir -p cmd/wego/flux/bin && touch cmd/wego/flux/bin/flux
+        run: mkdir -p pkg/flux/bin && touch pkg/flux/bin/flux
       - name: Run Snyk to check for vulnerabilities
         uses: snyk/actions/golang@master
         env:
@@ -61,7 +61,7 @@ jobs:
         with:
           sarif_file: snyk.code.sarif
       - name: Remove fake flux
-        run: rm -rv cmd/wego/flux/bin
+        run: rm -rv pkg/flux/bin
 
   snyk:
     name: Snyk
