+++ .github/workflows/test.yml
@@ -18,7 +18,7 @@ jobs:
     - name: Checkout code
       uses: actions/checkout@v2
     - name: Fake Install flux
-      run: mkdir -p cmd/wego/flux/bin && touch cmd/wego/flux/bin/flux
+      run: mkdir -p pkg/flux/bin && touch pkg/flux/bin/flux
     - name: Run linters
       uses: golangci/golangci-lint-action@v2
       with:
@@ -35,7 +35,7 @@ jobs:
     - name: Checkout code
       uses: actions/checkout@v2
     - name: Fake Install flux
-      run: mkdir -p cmd/wego/flux/bin && touch cmd/wego/flux/bin/flux
+      run: mkdir -p pkg/flux/bin && touch pkg/flux/bin/flux
     - name: Run tests
       run: make unit-tests
 
@@ -50,7 +50,7 @@ jobs:
     - name: Checkout code
       uses: actions/checkout@v2
     - name: Fake Install flux
-      run: mkdir -p cmd/wego/flux/bin && touch cmd/wego/flux/bin/flux
+      run: mkdir -p pkg/flux/bin && touch pkg/flux/bin/flux
     - name: Calc coverage
       run: |
         go test ./... -v -covermode=count -coverprofile=coverage.out
