+++ cmd/wego/main.go
@@ -9,6 +9,7 @@ import (
 	"github.com/spf13/cobra"
 	"github.com/weaveworks/weave-gitops/cmd/wego/flux"
 	"github.com/weaveworks/weave-gitops/cmd/wego/version"
+	fluxBin "github.com/weaveworks/weave-gitops/pkg/flux"
 )
 
 var options struct {
@@ -33,6 +34,7 @@ func configureLogger() {
 }
 
 func main() {
+	fluxBin.SetupFluxBin()
 	rootCmd.PersistentFlags().BoolVarP(&options.verbose, "verbose", "v", false, "Enable verbose output")
 	rootCmd.AddCommand(version.Cmd)
 	rootCmd.AddCommand(flux.Cmd)
