+++ tools/dependencies.toml
@@ -1,4 +1,4 @@
 [flux]
-version="0.11.0"
+version="0.12.3"
 tarpath="https://github.com/fluxcd/flux2/releases/download/v${version}/flux_${version}_${goos}_${goarch}.tar.gz"
-bindir="../cmd/wego/flux/bin"
+bindir="../pkg/flux/bin"
